api-server/requirements-quart.txt
# Quart Framework
quart>=0.19.0
quart-cors>=0.7.0
hypercorn>=0.16.0

# Flask-Security-Too (works with Quart)
flask-security-too>=5.3.0
bcrypt>=4.0.0

# Database
flask-sqlalchemy>=3.1.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# HTTP Client
httpx>=0.27.0

# Configuration & Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Validation (from shared libs)
# Install with: pip install -e ../shared/py_libs[all]

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code Quality
flake8>=7.0.0
black>=24.0.0
mypy>=1.8.0
